Refactor Preprocess class and add SpacyPreprocess class
